Timestamps are in float, 64 bit arith. seems messed up on Android. RBS seems to work functionally.

+++ wifi-rbs/jni/wifi-rbs.cpp
@@ -30,8 +30,8 @@ int main()  {
  /* now receive and log every packet */
  while(1) {
   Socket::Packet sync_pkt=wifi_socket.recv();
-  uint64_t ts=sync_pkt.timestamp;
-  fprintf(log_handle,"Received packet seq num %ld @ %lu \n",*(uint64_t *)sync_pkt.payload.data(),ts);  
+  long double ts=sync_pkt.timestamp; /* timestamp in secs */
+  fprintf(log_handle,"Received packet seq num %llu @ %lf \n",*(uint64_t *)sync_pkt.payload.data(),ts);  
  } 
  
  return 0;
